requirements.txt
asgiref==3.4.1
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
constantly==15.1.0
cryptography==35.0.0
cssselect==1.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0
django-active-link==0.1.8
django-allauth==0.45.0
django-appconf==1.0.5
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.4.0
django-embed-video==1.4.1
django-environ==0.8.1
django-humanize==0.1.2
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-watermark==0.2.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
humanize==4.0.0
hyperlink==21.0.0
idna==3.3
image==1.5.33
incremental==21.3.0
itemadapter==0.5.0
itemloaders==1.0.4
jmespath==1.0.0
lxml==4.8.0
oauthlib==3.1.1
parsel==1.6.0
Pillow==8.4.0
pip==22.0.3
Protego==0.2.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-decouple==3.6
python-memcached==1.59
python-resize-image==1.1.20
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.3
queuelib==1.6.2
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
Scrapy==2.6.1
service-identity==21.1.0
setuptools==57.0.0
six==1.16.0
social-auth-core==4.1.0
soupsieve==2.3.2
sqlparse==0.4.2
tldextract==3.2.1
Twisted==22.4.0
typing_extensions==4.1.1
urllib3==1.26.7
w3lib==1.22.0
wheel==0.36.2
whitenoise==5.3.0
zope.interface==5.4.0
